refactor: deleted service.ts and updated userService.ts
