+++ src/pages/SignInPage/SignInForm/index.tsx
@@ -96,7 +96,7 @@ function SignInForm() {
       dispatch(setUser(user));
       dispatch(setTokens(tokens));
 
-      navigate(urls.HOME_BASE);
+      navigate(urls.HOME);
     } catch (err) {
       const errorMessage = appErrors.FAILED_SIGN_IN;
