+++ src/pages/SignUpPage/SignUpForm/index.tsx
@@ -3,7 +3,7 @@ import { useForm, Controller, SubmitHandler } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 import { Typography } from '@mui/material';
 import { useAppDispatch } from 'src/redux/auth/hooks/hooks';
-import { useUserSignUpMutation } from 'src/redux/user/service';
+import { useUserSignUpMutation } from 'src/redux/user/userService';
 import { setUser } from 'src/redux/user/userSlice';
 import theme from 'src/theme';
 import { urls, validations } from 'src/common/constants';
@@ -92,9 +92,11 @@ function SignUpForm() {
           password,
         }).unwrap();
 
+        console.log(userData);
         dispatch(setUser(userData));
         navigate(urls.VERIFY);
       } catch (err) {
+        console.error(err);
         if (isFetchBaseQueryError(err) || isSerializedError(err)) {
           showToast('error', getErrorMessage(err));
         } else {
